server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-generativeai==0.8.0
python-multipart==0.0.6
pillow==10.1.0
python-dotenv==1.0.0
